Merge 02.11.2020

+++ cpp/OpenGL/Utils/GLFramebuffer.cpp
@@ -82,23 +82,23 @@ void GLFramebuffer::bind(const GLTexture2D& t0, const GLTexture2D& t1, const GLT
 void GLFramebuffer::bind(const GLTexture3D& t, size_t slice, const GLDepthBuffer& d) {
   GL(glBindFramebuffer(GL_FRAMEBUFFER, id));
   GL(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, d.getId()));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t.getId(), 0, slice));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t.getId(), 0, GLint(slice)));
   setBuffers(1,t.getWidth(), t.getHeight());
 }
 
 void GLFramebuffer::bind(const GLTexture3D& t0, size_t slice0, const GLTexture3D& t1, size_t slice1, const GLDepthBuffer& d) {
   GL(glBindFramebuffer(GL_FRAMEBUFFER, id));
   GL(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, d.getId()));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, slice0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, slice1));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, GLint(slice0)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, GLint(slice1)));
   setBuffers(2, t0.getWidth(), t0.getHeight());
 }
 void GLFramebuffer::bind(const GLTexture3D& t0, size_t slice0, const GLTexture3D& t1, size_t slice1, const GLTexture3D& t2, size_t slice2, const GLDepthBuffer& d) {
   GL(glBindFramebuffer(GL_FRAMEBUFFER, id));
   GL(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, d.getId()));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, slice0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, slice1));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT2, GL_TEXTURE_3D, t2.getId(), 0, slice2));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, GLint(slice0)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, GLint(slice1)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT2, GL_TEXTURE_3D, t2.getId(), 0, GLint(slice2)));
   setBuffers(3, t0.getWidth(), t0.getHeight());
 }
 
@@ -106,43 +106,43 @@ void GLFramebuffer::bind(const GLTexture3D& t0, size_t slice0, const GLTexture3D
                          const GLTexture3D& t3, size_t slice3, const GLDepthBuffer& d) {
   GL(glBindFramebuffer(GL_FRAMEBUFFER, id));
   GL(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, d.getId()));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, slice0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, slice1));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT2, GL_TEXTURE_3D, t2.getId(), 0, slice2));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT3, GL_TEXTURE_3D, t3.getId(), 0, slice3));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, GLint(slice0)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, GLint(slice1)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT2, GL_TEXTURE_3D, t2.getId(), 0, GLint(slice2)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT3, GL_TEXTURE_3D, t3.getId(), 0, GLint(slice3)));
   setBuffers(4, t0.getWidth(), t0.getHeight());
 }
 
 void GLFramebuffer::bind(const GLTexture3D& t, size_t slice) {
   GL(glBindFramebuffer(GL_FRAMEBUFFER, id));
   GL(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, 0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t.getId(), 0, slice));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t.getId(), 0, GLint(slice)));
   setBuffers(1,t.getWidth(), t.getHeight());
 }
 
 void GLFramebuffer::bind(const GLTexture3D& t0, size_t slice0, const GLTexture3D& t1, size_t slice1) {
   GL(glBindFramebuffer(GL_FRAMEBUFFER, id));
   GL(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, 0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, slice0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, slice1));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, GLint(slice0)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, GLint(slice1)));
   setBuffers(2, t0.getWidth(), t0.getHeight());
 }
 void GLFramebuffer::bind(const GLTexture3D& t0, size_t slice0, const GLTexture3D& t1, size_t slice1, const GLTexture3D& t2, size_t slice2) {
   GL(glBindFramebuffer(GL_FRAMEBUFFER, id));
   GL(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, 0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, slice0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, slice1));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT2, GL_TEXTURE_3D, t2.getId(), 0, slice2));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, GLint(slice0)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, GLint(slice1)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT2, GL_TEXTURE_3D, t2.getId(), 0, GLint(slice2)));
   setBuffers(3, t0.getWidth(), t0.getHeight());
 }
 
 void GLFramebuffer::bind(const GLTexture3D& t0, size_t slice0, const GLTexture3D& t1, size_t slice1, const GLTexture3D& t2, size_t slice2, const GLTexture3D& t3, size_t slice3) {
   GL(glBindFramebuffer(GL_FRAMEBUFFER, id));
   GL(glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, 0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, slice0));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, slice1));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT2, GL_TEXTURE_3D, t2.getId(), 0, slice2));
-  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT3, GL_TEXTURE_3D, t3.getId(), 0, slice3));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_3D, t0.getId(), 0, GLint(slice0)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_3D, t1.getId(), 0, GLint(slice1)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT2, GL_TEXTURE_3D, t2.getId(), 0, GLint(slice2)));
+  GL(glFramebufferTexture3D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT3, GL_TEXTURE_3D, t3.getId(), 0, GLint(slice3)));
   setBuffers(4, t0.getWidth(), t0.getHeight());
 }
 
@@ -198,5 +198,5 @@ void GLFramebuffer::setBuffers(size_t count, size_t width, size_t height) {
     }
   }
   
-  GL(glViewport(0, 0, width, height));
+  GL(glViewport(0, 0, GLint(width), GLint(height)));
 }
